move Experiment in separate module
