+++ openfl/component/experiment/__init__.py
@@ -1,0 +1,12 @@
+# Copyright (C) 2020-2021 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+"""Director package."""
+
+from .experiment import Experiment
+from .experiment import ExperimentsRegistry
+
+__all__ = [
+    'Experiment',
+    'ExperimentsRegistry',
+]
